requirements.txt
request 
colorama